Adds Monet.js
This is an ok library, but it doesn't support the Fantasy Land spec, so
it isn't really for me (why reinvent the wheel …). However, it is always
an inspiration to look at the source code of different implementations
and learn from them, so this could be added to this list.

This conflict brings up a question I have for other similar libraries:
should there maybe be a section called “Alternative Implementations” or
“Inspiration”? Older/less used libraries could go there.