Exploration: editing, additions, groupings
This is WIP and not ready to be published. Still wanted to share it for
feedback, if any :smile:

Dumped in some new stuff, libraries as well as resources. These need
proper descriptions and overlaps need to be removed (i.e. keep only the
resource that's the most awesome for a certain topic to make sure we
follow the Awesome list manifesto).

Experimented with grouping libraries into more specific categories, not
 really happy with it, will probably revert (at least the data
 structures one). We discussed this and @stoeffel's opinion was to keep
 a flat list of libraries. That might be the best idea after all. Will
 further experiment with the ordering of that list, however, maybe there
 is a better abstraction.